app/backend/pyproject.toml
[project]
name = "databricks-ads-voice-backend"
version = "0.1.0"
description = "FastAPI backend for voice-enabled Databricks ADS sessions"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "azure-ai-voicelive>=1.0.0b3",
    "azure-identity>=1.19.0",
    "github-copilot-sdk>=0.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
